+++ MK4duo/src/core/mechanics/core_mechanics.cpp
@@ -543,8 +543,30 @@ bool Core_Mechanics::position_is_reachable_by_probe(const float &rx, const float
       && WITHIN(ry, probe.min_y() - fslop, probe.max_y() + fslop);
 }
 
+// Report the real current position according to the steppers
+void Core_Mechanics::report_real_position() {
+
+  get_cartesian_from_steppers();
+
+  #if HAS_POSITION_MODIFIERS
+    xyze_pos_t npos = cartesian_position;
+    planner.unapply_modifiers(npos
+      #if HAS_LEVELING
+        , true
+      #endif
+    );
+  #else
+    const xyze_pos_t &npos = cartesian_position;
+  #endif
+
+  xyze_pos_t lpos = npos.asLogical();
+  lpos.e = planner.get_axis_position_mm(E_AXIS);
+  report_some_position(lpos);
+
+}
+
 // Report detail current position to host
-void Core_Mechanics::report_position_detail() {
+void Core_Mechanics::report_detail_position() {
 
   SERIAL_MSG("\nLogical:");
   report_xyz(position.asLogical());
@@ -584,6 +606,26 @@ void Core_Mechanics::report_position_detail() {
   const xyze_float_t diff = from_steppers - leveled;
   SERIAL_MSG("Differ: ");
   report_xyze(diff);
+
+}
+
+// Report the logical current position according to the most recent G-code command.
+void Core_Mechanics::report_logical_position() {
+
+  xyze_pos_t rpos = position;
+
+  #if HAS_POSITION_MODIFIERS
+    planner.apply_modifiers(rpos);
+  #endif
+
+  const abc_long_t spos = {
+    int32_t(LROUND(rpos.a * mechanics.steps_to_mm.a)),
+    int32_t(LROUND(rpos.b * mechanics.steps_to_mm.b)),
+    int32_t(LROUND(rpos.c * mechanics.steps_to_mm.c))
+  };
+
+  report_some_position(position.asLogical());
+  stepper.report_positions(spos);
 
 }